Move `ArenaInfo` to a separated file
